+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added functions for flattening 2D arrays to 1D arrays (#60)
 - Added functions to convert arrays to dictionaries (#60)
 - Added functions to find the last used row/column in 2D arrays (#75)
+- Added functions to find index tuple of a value inside a 2D array (#88)
 
 ### Added
 - Support for writing a 1D one-based array to a 2D one-based array (#9)